Refactor out table view
